feat: add WalletService.java file

+++ src/main/java/com/playtomic/tests/wallet/api/service/impl/WalletService.java
@@ -4,6 +4,6 @@ import com.playtomic.tests.wallet.api.dto.WalletDTO;
 
 public interface WalletService {
 
-    WalletDTO getWallet(long id);
+    WalletDTO getWallet(int id);
 
 }
